Clean and restructure the core context base.py (1000+ LoC for multiple concepts)
