+++ examples/core/heat_diffusion/heat_diffusion_cupy.py
@@ -37,7 +37,7 @@ import cupy as cp
 import matplotlib.pyplot as plt
 import numpy as np
 
-from pyfastflow.experimental.core.context.base import Bag
+from pyfastflow.experimental.core.context.bag import Bag
 from pyfastflow.experimental.core.context.cupy_backend import (
     CupyHelperBuilder,
     CupyKernelBuilder,
