+++ examples/core/heat_diffusion/heat_diffusion_routine_quadrants.py
@@ -42,7 +42,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import quadrants as qd
 
-from pyfastflow.experimental.core.context.base import Bag, merge
+from pyfastflow.experimental.core.context.bag import Bag, merge
 from pyfastflow.experimental.core.context.quadrants_backend import (
     QuadrantsHelperBuilder,
     QuadrantsKernelBuilder,
