+++ examples/core/heat_diffusion/heat_diffusion_routine_taichi.py
@@ -42,7 +42,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import taichi as ti
 
-from pyfastflow.experimental.core.context.base import Bag, merge
+from pyfastflow.experimental.core.context.bag import Bag, merge
 from pyfastflow.experimental.core.context.taichi_backend import (
     TaichiHelperBuilder,
     TaichiKernelBuilder,
