+++ examples/core/heat_diffusion/heat_diffusion_taichi.py
@@ -50,7 +50,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import taichi as ti
 
-from pyfastflow.experimental.core.context.base import Bag
+from pyfastflow.experimental.core.context.bag import Bag
 from pyfastflow.experimental.core.context.taichi_backend import (
     TaichiHelperBuilder,
     TaichiKernelBuilder,
