+++ examples/core/lem/lem_routine_cupy.py
@@ -55,7 +55,7 @@ import cupy as cp
 import matplotlib.pyplot as plt
 import numpy as np
 
-from pyfastflow.experimental.core.context.base import Bag, merge
+from pyfastflow.experimental.core.context.bag import Bag, merge
 from pyfastflow.experimental.core.context.cupy_backend import (
     CupyHelperBuilder,
     CupyKernelBuilder,
