+++ examples/core/lem/lem_routine_quadrants.py
@@ -45,7 +45,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import quadrants as qd
 
-from pyfastflow.experimental.core.context.base import Bag, merge
+from pyfastflow.experimental.core.context.bag import Bag, merge
 from pyfastflow.experimental.core.context.quadrants_backend import (
     QuadrantsHelperBuilder,
     QuadrantsKernelBuilder,
