+++ examples/core/shallow_water/shallow_water_taichi.py
@@ -37,7 +37,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 import taichi as ti
 
-from pyfastflow.experimental.core.context.base import Bag
+from pyfastflow.experimental.core.context.bag import Bag
 from pyfastflow.experimental.core.context.taichi_backend import (
     TaichiHelperBuilder,
     TaichiKernelBuilder,
